Removes Direct2D dependency from Game Objects (#3)
#Some code cleanup

#Fixed a bug that caused the game to reset when in the game over the state pressing buttons other than spacebar.

+++ src/GameClient/Game.h
@@ -19,13 +19,13 @@ namespace GameClient
 
 		std::chrono::steady_clock::time_point _lastUpdateTime = std::chrono::steady_clock::now();
 		std::chrono::nanoseconds _accumulatedFrameTime = std::chrono::nanoseconds(0);
-		unsigned short _maxUpdatesPerFrame = 1;
+		int _maxUpdatesPerFrame = 1;
 		std::chrono::milliseconds _timeStep = std::chrono::milliseconds(_normal);
 
 	public:
 		explicit Game(std::unique_ptr<Engine::IEngine> engine);
 		~Game();
-		void OnResize(UINT width, UINT height) const;
+		void OnResize(int width, int height) const;
 		void OnInput(Input::Input input);
 
 		Game(Game& copyOther) = delete;
